Move most of history into terminalContrib
Part of #230129
