+++ src/vs/workbench/contrib/terminalContrib/history/browser/terminalRunRecentQuickPick.ts
@@ -3,31 +3,32 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Toggle } from '../../../../base/browser/ui/toggle/toggle.js';
-import { isMacintosh, OperatingSystem } from '../../../../base/common/platform.js';
-import { ITextModel } from '../../../../editor/common/model.js';
-import { IModelService } from '../../../../editor/common/services/model.js';
-import { ITextModelContentProvider, ITextModelService } from '../../../../editor/common/services/resolverService.js';
-import { localize } from '../../../../nls.js';
-import { IInstantiationService, ServicesAccessor } from '../../../../platform/instantiation/common/instantiation.js';
-import { IQuickInputButton, IQuickInputService, IQuickPickItem, IQuickPickSeparator } from '../../../../platform/quickinput/common/quickInput.js';
-import { ITerminalCommand, TerminalCapability } from '../../../../platform/terminal/common/capabilities/capabilities.js';
-import { collapseTildePath } from '../../../../platform/terminal/common/terminalEnvironment.js';
-import { asCssVariable, inputActiveOptionBackground, inputActiveOptionBorder, inputActiveOptionForeground } from '../../../../platform/theme/common/colorRegistry.js';
-import { ThemeIcon } from '../../../../base/common/themables.js';
-import { ITerminalInstance } from './terminal.js';
-import { commandHistoryFuzzySearchIcon, commandHistoryOutputIcon, commandHistoryRemoveIcon } from './terminalIcons.js';
-import { getCommandHistory, getDirectoryHistory, getShellFileHistory } from '../common/history.js';
-import { TerminalStorageKeys } from '../common/terminalStorageKeys.js';
-import { terminalStrings } from '../common/terminalStrings.js';
-import { URI } from '../../../../base/common/uri.js';
-import { fromNow } from '../../../../base/common/date.js';
-import { IEditorService } from '../../../services/editor/common/editorService.js';
-import { showWithPinnedItems } from '../../../../platform/quickinput/browser/quickPickPin.js';
-import { IStorageService } from '../../../../platform/storage/common/storage.js';
-import { IContextKey } from '../../../../platform/contextkey/common/contextkey.js';
-import { AccessibleViewProviderId, IAccessibleViewService } from '../../../../platform/accessibility/browser/accessibleView.js';
-import { Disposable, DisposableStore } from '../../../../base/common/lifecycle.js';
+import { Toggle } from '../../../../../base/browser/ui/toggle/toggle.js';
+import { isMacintosh, OperatingSystem } from '../../../../../base/common/platform.js';
+import { ITextModel } from '../../../../../editor/common/model.js';
+import { IModelService } from '../../../../../editor/common/services/model.js';
+import { ITextModelContentProvider, ITextModelService } from '../../../../../editor/common/services/resolverService.js';
+import { localize } from '../../../../../nls.js';
+import { IInstantiationService, ServicesAccessor } from '../../../../../platform/instantiation/common/instantiation.js';
+import { IQuickInputButton, IQuickInputService, IQuickPickItem, IQuickPickSeparator } from '../../../../../platform/quickinput/common/quickInput.js';
+import { ITerminalCommand, TerminalCapability } from '../../../../../platform/terminal/common/capabilities/capabilities.js';
+import { collapseTildePath } from '../../../../../platform/terminal/common/terminalEnvironment.js';
+import { asCssVariable, inputActiveOptionBackground, inputActiveOptionBorder, inputActiveOptionForeground } from '../../../../../platform/theme/common/colorRegistry.js';
+import { ThemeIcon } from '../../../../../base/common/themables.js';
+import { ITerminalInstance } from '../../../terminal/browser/terminal.js';
+import { commandHistoryFuzzySearchIcon, commandHistoryOutputIcon, commandHistoryRemoveIcon } from '../../../terminal/browser/terminalIcons.js';
+import { getDirectoryHistory } from '../../../terminal/common/history.js';
+import { TerminalStorageKeys } from '../../../terminal/common/terminalStorageKeys.js';
+import { terminalStrings } from '../../../terminal/common/terminalStrings.js';
+import { URI } from '../../../../../base/common/uri.js';
+import { fromNow } from '../../../../../base/common/date.js';
+import { IEditorService } from '../../../../services/editor/common/editorService.js';
+import { showWithPinnedItems } from '../../../../../platform/quickinput/browser/quickPickPin.js';
+import { IStorageService } from '../../../../../platform/storage/common/storage.js';
+import { IContextKey } from '../../../../../platform/contextkey/common/contextkey.js';
+import { AccessibleViewProviderId, IAccessibleViewService } from '../../../../../platform/accessibility/browser/accessibleView.js';
+import { Disposable, DisposableStore } from '../../../../../base/common/lifecycle.js';
+import { getCommandHistory, getShellFileHistory } from '../common/history.js';
 
 export async function showRunRecentQuickPick(
 	accessor: ServicesAccessor,
