+++ src/vs/workbench/contrib/terminalContrib/history/test/common/history.test.ts
@@ -4,22 +4,23 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { deepStrictEqual, strictEqual, ok } from 'assert';
-import { VSBuffer } from '../../../../../base/common/buffer.js';
-import { Schemas } from '../../../../../base/common/network.js';
-import { join } from '../../../../../base/common/path.js';
-import { isWindows, OperatingSystem } from '../../../../../base/common/platform.js';
-import { env } from '../../../../../base/common/process.js';
-import { URI } from '../../../../../base/common/uri.js';
-import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../base/test/common/utils.js';
-import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
-import { TestConfigurationService } from '../../../../../platform/configuration/test/common/testConfigurationService.js';
-import { IFileService } from '../../../../../platform/files/common/files.js';
-import { TestInstantiationService } from '../../../../../platform/instantiation/test/common/instantiationServiceMock.js';
-import { IRemoteAgentEnvironment } from '../../../../../platform/remote/common/remoteAgentEnvironment.js';
-import { IStorageService } from '../../../../../platform/storage/common/storage.js';
-import { fetchBashHistory, fetchFishHistory, fetchPwshHistory, fetchZshHistory, ITerminalPersistedHistory, sanitizeFishHistoryCmd, TerminalPersistedHistory } from '../../common/history.js';
-import { IRemoteAgentConnection, IRemoteAgentService } from '../../../../services/remote/common/remoteAgentService.js';
-import { TestStorageService } from '../../../../test/common/workbenchTestServices.js';
+import { VSBuffer } from '../../../../../../base/common/buffer.js';
+import { Schemas } from '../../../../../../base/common/network.js';
+import { join } from '../../../../../../base/common/path.js';
+import { isWindows, OperatingSystem } from '../../../../../../base/common/platform.js';
+import { env } from '../../../../../../base/common/process.js';
+import { URI } from '../../../../../../base/common/uri.js';
+import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../../../base/test/common/utils.js';
+import { IConfigurationService } from '../../../../../../platform/configuration/common/configuration.js';
+import { TestConfigurationService } from '../../../../../../platform/configuration/test/common/testConfigurationService.js';
+import { IFileService } from '../../../../../../platform/files/common/files.js';
+import { TestInstantiationService } from '../../../../../../platform/instantiation/test/common/instantiationServiceMock.js';
+import { IRemoteAgentEnvironment } from '../../../../../../platform/remote/common/remoteAgentEnvironment.js';
+import { IStorageService } from '../../../../../../platform/storage/common/storage.js';
+import { ITerminalPersistedHistory, TerminalPersistedHistory } from '../../../../terminal/common/history.js';
+import { IRemoteAgentConnection, IRemoteAgentService } from '../../../../../services/remote/common/remoteAgentService.js';
+import { TestStorageService } from '../../../../../test/common/workbenchTestServices.js';
+import { fetchBashHistory, fetchFishHistory, fetchPwshHistory, fetchZshHistory, sanitizeFishHistoryCmd } from '../../common/history.js';
 
 function getConfig(limit: number) {
 	return {
